sample 2/requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
tqdm==4.66.1
python-multipart==0.0.6
gunicorn==21.2.0